core/requirements.txt
aiofiles==0.6.0
aiounittest==1.4.0
appdirs==1.4.4
bcrypt==3.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
distlib==0.3.1
dnspython==2.1.0
ecdsa==0.14.1
fastapi==0.63.0
fastapi-utils==0.2.1
filelock==3.0.12
gcloud==0.18.3
google-api-core==1.26.2
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-secret-manager==2.3.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
gunicorn==20.0.4
h11==0.12.0
httpcore==0.12.3
httplib2==0.19.0
httptools==0.1.1
httpx==0.17.1
idna==2.10
importlib-metadata==3.7.3
kiwisolver==1.3.1
libcst==0.3.17
matplotlib==3.3.4
mypy-extensions==0.4.3
numpy==1.20.1
oauth2client==4.1.3
packaging==20.9
pandas==1.2.3
passlib==1.7.4
Pillow==8.1.2
pipenv==2020.11.15
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
scipy==1.6.2
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.2
starlette==0.13.6
typer==0.3.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.4.1
